requirements.txt
# D:\cekfakta-api\requirements.txt

google-generativeai